requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
filelock==3.18.0
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
fsspec==2025.3.2
-e git+https://github.com/philippweder/hmpinn.git@2db42439187e565f0b2ae7486e6f713601f83e06#egg=hmpinn
hydra-core==1.3.2
Jinja2==3.1.6
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
llvmlite @ file:///C:/b/abs_da51wefv9a/croot/llvmlite_1741209883316/work
MarkupSafe==3.0.2
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath==1.3.0
networkx==3.4.2
numba @ file:///D:/bld/numba_1744232058805/work
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
nutils==9.0
nutils-poly==1.0.1
omegaconf==2.3.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp310-cp310-win_amd64.whl#sha256=6fc0a0523ea4ec3f23a2b790c738c857942b4c579479f8012ac8148d675b8975
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt5==5.15.9
PyQt5-sip @ file:///D:/bld/pyqt-split_1695418047530/work/pyqt_sip
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
PyYAML==6.0.2
seaborn @ file:///C:/b/abs_522_0_dy9p/croot/seaborn_1741187268016/work
sip @ file:///D:/bld/sip_1697300537181/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
stringly==1.0b3
sympy==1.13.1
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
treelog==1.0
typing_extensions==4.13.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
